suraksha_setu/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-huggingface
langchain-groq
langchain-chroma
chromadb
sentence-transformers
pypdf
pydantic
numpy
faiss-cpu